Rename methods in BackStackNavigator, move BackStackNavigator to :common_navigation_compose module
