+++ modules/app_ui_kit_demo/src/main/kotlin/kekmech/ru/mpeiapp/demo/BackStackRootNode.kt
@@ -10,8 +10,8 @@ import com.bumble.appyx.core.node.Node
 import com.bumble.appyx.core.node.ParentNode
 import com.bumble.appyx.navmodel.backstack.BackStack
 import com.bumble.appyx.navmodel.backstack.transitionhandler.rememberBackstackFader
-import kekmech.ru.common_navigation_api.BackStackNavigator
 import kekmech.ru.common_navigation_api.NavTarget
+import kekmech.ru.common_navigation_compose.BackStackNavigator
 import kekmech.ru.common_navigation_compose.LocalBackStackNavigator
 import kekmech.ru.mpeiapp.demo.screens.main.MainScreenNavTarget
