+++ modules/app_ui_kit_demo/src/main/kotlin/kekmech/ru/mpeiapp/demo/screens/colors/ColorsScreen.kt
@@ -45,7 +45,7 @@ private fun ColorsScreen() {
                     scrolledContainerColor = MpeixTheme.palette.surface,
                 ),
                 navigationIcon = {
-                    IconButton(onClick = { navigator.pop() }) {
+                    IconButton(onClick = { navigator.back() }) {
                         Icon(
                             imageVector = Icons.Default.ArrowBack,
                             contentDescription = null,
