+++ modules/app_ui_kit_demo/src/main/kotlin/kekmech/ru/mpeiapp/demo/screens/typography/TypographyScreen.kt
@@ -46,7 +46,7 @@ private fun TypographyScreen() {
                     scrolledContainerColor = MpeixTheme.palette.surface,
                 ),
                 navigationIcon = {
-                    IconButton(onClick = { navigator.pop() }) {
+                    IconButton(onClick = { navigator.back() }) {
                         Icon(
                             imageVector = Icons.Default.ArrowBack,
                             contentDescription = null,
