+++ modules/common/navigation/api/src/main/kotlin/kekmech/ru/common_navigation_api/NavTarget.kt
@@ -4,6 +4,20 @@ import android.os.Parcelable
 import com.bumble.appyx.core.modality.BuildContext
 import com.bumble.appyx.core.node.Node
 
+/**
+ * Navigation target for global backstack model navigation.
+ * Implement this interface to create a screen that can be navigated to inside the backstack.
+ *
+ * Usage:
+ * ```kotlin
+ * @Parcelize
+ * internal class SomeScreenNavTarget(private val arg: SomeArg) : NavTarget {
+ *
+ *     override fun resolve(buildContext: BuildContext): Node =
+ *         node(buildContext) { SomeScreen(arg) }
+ * }
+ * ```
+ */
 interface NavTarget : Parcelable {
 
     fun resolve(buildContext: BuildContext): Node
