+++ modules/common/navigation/compose/build.gradle.kts
@@ -5,4 +5,6 @@ plugins {
 
 dependencies {
     api(project(":common_navigation_api"))
+
+    implementation(libs.appyx.core)
 }
