+++ modules/common/navigation/compose/src/main/kotlin/kekmech/ru/common_navigation_compose/LocalBackStackNavigator.kt
@@ -1,7 +1,6 @@
 package kekmech.ru.common_navigation_compose
 
 import androidx.compose.runtime.staticCompositionLocalOf
-import kekmech.ru.common_navigation_api.BackStackNavigator
 
 val LocalBackStackNavigator = staticCompositionLocalOf<BackStackNavigator> {
     error("StaticCompositionalLocal BackStackNavigator not provided")
